Remove BasicQuestHelper usage for QuestHelper usage
Allows for other QuestHelper types to be more easily used in the future.
